fix: challengeAdmin - update

+++ src/main/java/com/posmosalimos/geulgwi/api/user/update/controller/UpdateController.java
@@ -39,10 +39,10 @@ public class UpdateController {
 
         User user = updateService.update(userSeq, updateDTO);
 
-        if (fileService.findByUser(user) != null && !file.isEmpty()) {
+        if (fileService.findByUser(user) != null && file != null && !file.isEmpty()) {
             fileService.removeUserFile(user);
             fileService.storeUserFile(user, file);
-        } else if (file.isEmpty())
+        } else if (file != null && !file.isEmpty())
             fileService.removeUserFile(user);
 
         log.info("user - update success(userSeq: {})", userSeq);
